+++ src/strassen.cu
@@ -194,9 +194,9 @@ void strassen(Float *A, Float *B, Float *C,
   bool stop = false;
   
 #if 1
-  float mm = CUTOFF / XB2;
-  float nn = CUTOFF / YA2;
-  float kk = CUTOFF / XA2;
+  float mm = (float)CUTOFF / XB2;
+  float nn = (float)CUTOFF / YA2;
+  float kk = (float)CUTOFF / XA2;
   if ((mm + nn + kk) >= 3) {
       stop = true;
   }
